build view
